+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/aip/reference/contributor/AipReferenceContributor.kt
@@ -12,6 +12,7 @@ import io.kanro.idea.plugin.protobuf.aip.AipOptions
 import io.kanro.idea.plugin.protobuf.aip.reference.AipResourceReference
 import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufRpcInputFieldReference
 import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufRpcOutputFieldReference
+import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufTypeNameInStringReference
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFieldDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufOptionAssign
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufRpcDefinition
@@ -19,6 +20,8 @@ import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufStringValue
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.stratify.ProtobufOptionHover
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure.ProtobufValueAssign
 import io.kanro.idea.plugin.protobuf.lang.psi.value
+import io.kanro.idea.plugin.protobuf.string.splitToRange
+import io.kanro.idea.plugin.protobuf.string.trim
 
 class AipReferenceContributor : PsiReferenceContributor() {
     override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
@@ -34,6 +37,13 @@ class AipReferenceContributor : PsiReferenceContributor() {
                 .inside(ProtobufRpcDefinition::class.java)
                 .inside(ProtobufOptionAssign::class.java),
             AipFieldReferenceProvider()
+        )
+
+        registrar.registerReferenceProvider(
+            PlatformPatterns.psiElement(ProtobufStringValue::class.java)
+                .inside(ProtobufRpcDefinition::class.java)
+                .inside(ProtobufOptionAssign::class.java),
+            AipTypeNameReferenceProvider()
         )
     }
 }
@@ -68,5 +78,29 @@ class AipFieldReferenceProvider : PsiReferenceProvider() {
         if (targetField == AipOptions.httpRuleBodyName) return ProtobufRpcInputFieldReference(element)
         if (targetField == AipOptions.httpRuleResponseBodyName) return ProtobufRpcOutputFieldReference(element)
         return null
+    }
+}
+
+class AipTypeNameReferenceProvider : PsiReferenceProvider() {
+    override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<PsiReference> {
+        if (element !is ProtobufStringValue) return arrayOf()
+        val assign = element.parentOfType<ProtobufValueAssign>() ?: return arrayOf()
+        val targetField = assign.field()?.qualifiedName()
+        if (targetField == AipOptions.lroMetadataName) return getReference(element)
+        if (targetField == AipOptions.lroResponseName) return getReference(element)
+        return arrayOf()
+    }
+
+    private fun getReference(element: ProtobufStringValue): Array<PsiReference> {
+        val text = element.text
+        val ranges = text.splitToRange('.', true)
+        var reference: ProtobufTypeNameInStringReference? = null
+        return ranges.asReversed().map {
+            ProtobufTypeNameInStringReference(
+                element, it.trim(text, '"'), reference
+            ).apply {
+                reference = this
+            }
+        }.toTypedArray()
     }
 }